+++ src/main/java/roomescape/payment/domain/PaymentClient.java
@@ -1,7 +1,7 @@
-package roomescape.reservation.infrastructure;
+package roomescape.payment.domain;
 
 import roomescape.reservation.presentation.dto.PaymentRequest;
-import roomescape.reservation.presentation.dto.PaymentResponse;
+import roomescape.payment.infrastructure.dto.PaymentResponse;
 
 public interface PaymentClient {
