+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -11,6 +11,7 @@ import roomescape.auth.login.presentation.dto.SearchCondition;
 import roomescape.common.util.time.DateTime;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.MemberRepository;
+import roomescape.payment.service.PaymentService;
 import roomescape.reservation.infrastructure.dto.WaitingWithRank;
 import roomescape.member.exception.MemberNotFound;
 import roomescape.member.presentation.dto.MemberResponse;
