+++ src/test/java/roomescape/reservation/service/ReservationServiceTest.java
@@ -25,6 +25,7 @@ import roomescape.member.domain.MemberRepository;
 import roomescape.member.infrastructure.JpaMemberRepository;
 import roomescape.member.infrastructure.JpaMemberRepositoryAdapter;
 import roomescape.member.presentation.dto.MyReservationResponse;
+import roomescape.payment.service.PaymentService;
 import roomescape.reservation.domain.ReservationRepository;
 import roomescape.reservation.domain.WaitingRepository;
 import roomescape.reservation.exception.ReservationException;
@@ -32,8 +33,8 @@ import roomescape.reservation.infrastructure.JpaReservationRepository;
 import roomescape.reservation.infrastructure.JpaReservationRepositoryAdapter;
 import roomescape.reservation.infrastructure.JpaWaitingRepository;
 import roomescape.reservation.infrastructure.JpaWaitingRepositoryAdapter;
-import roomescape.reservation.infrastructure.PaymentClient;
-import roomescape.reservation.infrastructure.TossPaymentClient;
+import roomescape.payment.domain.PaymentClient;
+import roomescape.payment.infrastructure.TossPaymentClient;
 import roomescape.reservation.presentation.dto.ReservationRequest;
 import roomescape.reservation.presentation.dto.ReservationResponse;
 import roomescape.reservation.presentation.dto.WaitingResponse;
